feat: 评测体系补全——collect 模式与首轮人工抽检结论
- evaluate.py 新增 collect: 抽检帧(带框图+检测JSON)与红绿灯裁剪图采集
- 修复 detection_engine 输出 numpy 类型不可 JSON 序列化
- 首轮人工抽检 12帧/15框: 精确率约13%，行人→car、前景手套→bicycle 等 OOD 误检集中
- 归因为录屏素材与训练数据域差距(竖屏letterbox/UI/第一人称前景)，非系统缺陷
- 依据证据将 confidence 0.5→0.6(宁漏勿误)；报告含完整方法学与5条改进清单
- 全片无 traffic_light 检出，灯色真实评测留待实拍样本(如实记录)

+++ core/detection_engine.py
@@ -158,12 +158,12 @@ class DetectionEngine:
                 confidence = float(box.conf[0].cpu().numpy())
                 class_name = self.class_names.get(class_id, f'class_{class_id}')
 
-                # 计算中心点和面积
-                center_x = (x1 + x2) / 2
-                center_y = (y1 + y2) / 2
-                width = x2 - x1
-                height = y2 - y1
-                area = width * height
+                # 计算中心点和面积（统一转为 Python 原生类型，保证可 JSON 序列化）
+                center_x = float((x1 + x2) / 2)
+                center_y = float((y1 + y2) / 2)
+                width = int(x2 - x1)
+                height = int(y2 - y1)
+                area = int(width * height)
 
                 # 边界检查
                 x1 = max(0, min(x1, frame_w))
